feature:Artifactory支持文件托管服务 #2000
类似方法抽象
